Add delay for deleting duplicate ROIs
There seems to be some latency between this script and the Omero server
that results in adding a new roi instead replacing the duplicate geometry.
Ezomero doesn't see allow changing the label or colour of an existing
ROI, hence the need to deleta and add new. Duplicates are still sometimes
added though.